+++ feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/viewmodel/QuizViewModel.kt
@@ -1,0 +1,8 @@
+package kr.boostcamp_2024.course.quiz.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+
+@HiltViewModel
+class QuizViewModel : ViewModel() {
+}
